download exercise images, wip
