+++ dockerfiles/steps/git-assemble.bash
@@ -23,7 +23,9 @@ while read -r line; do # Loop over each <book> entry in the META-INF/books.xml m
 
     try neb assemble "$IO_FETCH_META/modules" temp-assembly/
 
-    try cp "temp-assembly/collection.assembled.xhtml" "$IO_ASSEMBLED/$slug.assembled.xhtml"
+    ## download exercise images and replace internet links with local resource links
+    try download-exercise-images "../resources" "temp-assembly/collection.assembled.xhtml" "$IO_ASSEMBLED/$slug.assembled.xhtml"
+
     try rm -rf temp-assembly
     try rm "$IO_FETCH_META/modules/collection.xml"
